requirements.txt
streamlit_lottie
matplotlib.pyplot
opencv-python
matplotlib
requests
pandas
plost
altair
openpyxl
plotly
dash
json
plotly.express
numpy
tkinter
multiapp
seaborn
streamlit_option_menu
turtle
cProfile  
curses  
tarfile
tensorflow
sqlite3
requests

